ffi: Move more things into ruma module
